Multiple textures now working
Previously, the method of pushing textures to a shape was not working. Textures were being saved as values, not references, when pushed to the vector. This was apparently causing the textures to be destroyed. Only the most recently pushed texture was being drawn for some reason. The vector of textures in the shape object has been changed to store pointers to vectors, so now multiple textures can be referenced properly. The user will be responsible for managing the memory of their texture objects.